requirements.txt
flask>=3.0
werkzeug>=3.0
Pillow>=10.0
pdf2image>=1.17
python-pptx>=0.6
